requirements.txt
uvicorn[standard]
fastapi==0.133.1
passlib==1.7.4
pydantic==2.12.5
PyJWT==2.11.0
python-dotenv==1.2.1
SQLAlchemy==2.0.47
python-multipart
pydantic[email]
argon2_cffi
groq~=1.0.0
transformers~=5.2.0
requests~=2.32.5